+++ dime/lib/dime.dart
@@ -5,5 +5,6 @@ library dime;
 
 export 'src/common.dart';
 export 'src/dime_base.dart';
+export 'src/dime_module.dart';
 export 'src/dime_static.dart';
 export 'src/factories.dart';