+++ dime/lib/src/factories.dart
@@ -1,118 +1,11 @@
 import 'package:dime/dime.dart';
 import 'package:dime/src/common.dart';
-import 'package:fimber/fimber.dart';
 
-/// Base Dime module for providing types and [InjectFactory] for types.
-abstract class BaseDimeModule with Closable {
-  final Map<Type, InjectFactory> _injectMap = {};
 
-  /// Returns module's injection map.
-  Map<Type, InjectFactory> get injectMap => _injectMap;
+/// Typed type to create a instance with a [tag] definition provided.
+/// This type is used in generative InjectFactory
+typedef Creator<T> = T Function(String tag);
 
-  /// Updates all injection - abstract method
-  /// to be implemented by main [BaseDimeModule]
-  void updateInjections();
-
-  /// Adds a [factory] into module's injection map.
-  void addFactory(Type type, InjectFactory factory) {
-    _injectMap[type] = factory;
-  }
-
-  /// Removes a [type] from module's injection map.
-  void remove(Type type) {
-    _injectMap.remove(type);
-  }
-
-  /// Adds [InjectFactory] with [objectCreator] function.
-  void addCreator<T>(Creator<T> objectCreator) {
-    injectMap[T] = CreatorInjectFactory<T>(objectCreator);
-  }
-
-  /// Adds singleton [InjectFactory] with [objectCreator] function
-  void addSingleByCreator<T>(Creator<T> objectCreator) {
-    injectMap[T] = CreatorSingleInjectFactory<T>(objectCreator);
-  }
-
-  /// Will add instance to the module, if T is defined in <T> method call,
-  /// it will find the instance faster.
-  /// Otherwise it would find first instance that matches type check with T
-  void addSingle<T>(T instance, {String tag}) {
-    if (instance is T) {
-      if (tag == null) {
-        if (injectMap.containsKey(T)) {
-          Closable.closeWith(injectMap[T]);
-        }
-        injectMap[T] = SingleInstanceFactory<T>(instance);
-      } else {
-        var instanceFactory = injectMap[T];
-        if (instanceFactory == null) {
-          instanceFactory = SingleByTagInstanceFactory<T>();
-          injectMap[T] = instanceFactory;
-        }
-        if (instanceFactory is SingleInjectFactory<T>) {
-          // remove from single ad make it taggable factory
-          var oldSingleInjectFactory =
-              (instanceFactory as SingleInjectFactory<T>);
-          instanceFactory = SingleByTagInstanceFactory<T>();
-          _injectMap[T] = instanceFactory;
-          if (oldSingleInjectFactory._localSingleton != null) {
-            (instanceFactory as SingleByTagInstanceFactory<T>).put(
-                InjectTagFactory.defaultTag,
-                oldSingleInjectFactory._localSingleton);
-          }
-          (instanceFactory as SingleByTagInstanceFactory<T>).put(tag, instance);
-        }
-        if (instanceFactory is SingleByTagInstanceFactory<T>) {
-          instanceFactory.put(tag, instance);
-        }
-      }
-    } else {
-      throw DimeException.message("Instance provided does not match type.");
-    }
-  }
-
-  /// Gets the created value from the module
-  T get<T>({String tag}) {
-    var name = T.toString();
-    var injectFactory = injectMap[T];
-
-    if (injectFactory != null && injectFactory is InjectTagFactory) {
-      // use default tag for TaggedInjectFactory
-      var instance =
-          injectFactory.createTagged(tag ?? InjectTagFactory.defaultTag);
-      Fimber.d("Injecting: $name for tag $tag with $instance");
-      return instance;
-    } else if (injectFactory != null && injectFactory is InjectFactory) {
-      if (tag != null && tag != InjectTagFactory.defaultTag) {
-        return null; // Dime not provide providing instance,
-        // because tagged instance was expected.
-      } else {
-        var instance = injectFactory.create();
-        Fimber.d("Injecting: $name with $instance");
-        return instance;
-      }
-    } else {
-      return null;
-    }
-  }
-
-  /// Injects the created value from the module
-  /// [Deprecated] - use [get]
-  @deprecated
-  T inject<T>({String tag}) {
-    return get(tag: tag);
-  }
-
-  /// Closes the module and all its [InjectFactory].
-  @override
-  void close() {
-    _injectMap.forEach((type, factory) {
-      if (factory is SingleInjectFactory) {
-        Closable.closeWith(factory);
-      }
-    });
-  }
-}
 
 /// Singleton [InjectFactory] for a [T] type.
 /// It wil create an object using [Creator] method
@@ -153,10 +46,6 @@ class CreatorInjectFactory<T> extends InjectTagFactory<T> {
   }
 }
 
-/// Typed type to create a instance with a [tag] definition provided.
-/// This type is used in generative InjectFactory
-typedef Creator<T> = T Function(String tag);
-
 /// InjectFactory that stores singleton instances per tag.
 class SingleByTagInstanceFactory<T> extends TaggedSingletonInjectFactory<T> {
   /// Adds [instance] for a given [tag], overriding old value if exists.
@@ -183,9 +72,7 @@ class SingleByTagInstanceFactory<T> extends TaggedSingletonInjectFactory<T> {
 /// Single instance factory where instance created is given at constructor.
 class SingleInstanceFactory<T> extends SingleInjectFactory<T> {
   /// Creates [InjectFactory] with single [instance] of [T] type.
-  SingleInstanceFactory(T instance) {
-    _localSingleton = instance;
-  }
+  SingleInstanceFactory(T instance):super(instance);
 
   /// Implementation of [createInstance] should never be called
   @override
@@ -197,7 +84,11 @@ class SingleInstanceFactory<T> extends SingleInjectFactory<T> {
 /// of the T object.
 abstract class SingleInjectFactory<T> extends InjectFactory<T> with Closable {
   T _localSingleton;
-
+  /// Local singleton value created in this factory
+  T get localSingleton => _localSingleton;
+  /// Creates instance of SingleInjectFactory 
+  /// with preloaded instance as optional.
+  SingleInjectFactory([this._localSingleton]);
   /// Creates instance.
   T createInstance();